Move COntexts to a specific folder
