+++ react/javascript/src/components/gherkin/ExamplesTableBody.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { messages } from '@cucumber/messages'
-import CucumberQueryContext from '../../CucumberQueryContext'
+import CucumberQueryContext from '../../contexts/CucumberQueryContext'
 import isNumber from './isNumber'
-import GherkinQueryContext from '../../GherkinQueryContext'
+import GherkinQueryContext from '../../contexts/GherkinQueryContext'
 import ErrorMessage from './ErrorMessage'
 import statusName from './statusName'
 import StatusIcon from './StatusIcon'
