+++ react/javascript/src/components/gherkin/Step.tsx
@@ -3,8 +3,8 @@ import DataTable from './DataTable'
 import Keyword from './Keyword'
 import DocString from './DocString'
 import { messages } from '@cucumber/messages'
-import CucumberQueryContext from '../../CucumberQueryContext'
-import GherkinQueryContext from '../../GherkinQueryContext'
+import CucumberQueryContext from '../../contexts/CucumberQueryContext'
+import GherkinQueryContext from '../../contexts/GherkinQueryContext'
 import ErrorMessage from './ErrorMessage'
 import StepContainer from './StepContainer'
 import Attachment from './Attachment'
