+++ react/javascript/src/contexts/IdGeneratorContext.ts
@@ -1,4 +1,4 @@
 import React from 'react'
-import IdGenerator from './IdGenerator'
+import IdGenerator from '../IdGenerator'
 
 export default React.createContext(new IdGenerator())
