+++ react/javascript/src/index.ts
@@ -2,10 +2,12 @@ import GherkinDocumentList from './components/app/GherkinDocumentList'
 import FilteredResults from './components/app/FilteredResults'
 import QueriesWrapper from './components/app/QueriesWrapper'
 import filterByStatus from './filter/filterByStatus'
-import GherkinQueryContext from './GherkinQueryContext'
-import CucumberQueryContext from './CucumberQueryContext'
-import SearchQueryContext from './SearchQueryContext'
-import EnvelopesQueryContext, { EnvelopesQuery } from './EnvelopesQueryContext'
+import GherkinQueryContext from './contexts/GherkinQueryContext'
+import CucumberQueryContext from './contexts/CucumberQueryContext'
+import SearchQueryContext from './contexts/SearchQueryContext'
+import EnvelopesQueryContext, {
+  EnvelopesQuery,
+} from './contexts/EnvelopesQueryContext'
 
 export {
   GherkinDocumentList,
