+++ react/javascript/test/messageTests.tsx
@@ -7,7 +7,7 @@ import { JSDOM } from 'jsdom'
 import { QueriesWrapper } from '../src'
 import { Query as CucumberQuery } from '@cucumber/query'
 import GherkinDocumentList from '../src/components/app/GherkinDocumentList'
-import { EnvelopesQuery } from '../src/EnvelopesQueryContext'
+import { EnvelopesQuery } from '../src/contexts/EnvelopesQueryContext'
 import { promisify } from 'util'
 import { pipeline, Writable } from 'stream'
 import glob from 'glob'
